move sample_rate tests to PointsetSequences.java and skip if no pointset configured (#591)
